refactor: actions 컨벤션 통일
