+++ src/features/(authenticated)/admin/root/components/track/TrackCreate.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import TrackFormBase from '@/features/(authenticated)/admin/root/components/track/TrackFormBase'
-import { createTrackAction } from '@/features/(authenticated)/admin/track/actions/createTrack'
+import { createTrackAction } from '@/features/(authenticated)/admin/track/actions/createTrack.action'
 import { initialState } from '@/features/(authenticated)/admin/root/components/track/TrackFiled'
 
 export default function TrackCreate() {
